requirements.txt
flask
openai
Pillow
flask-cors